requirements.txt
flask
tensorflow-cpu
gunicorn
numpy
Pillow